requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
bootstrap4==0.1.0
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.1.5
django-bootstrap4==2.3.1
django-crispy-forms==1.10.0
django-heroku==0.3.1
django-rest-passwordreset==1.1.0
django-tinymce==3.2.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
importlib-metadata==2.1.1
psycopg2==2.8.6
python-dateutil==2.8.1
python-decouple==3.4
pytz==2020.5
pyuploadcare==2.7.0
requests==2.25.1
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
zipp==3.4.0
